feat: Separate Detail and Simple ticket
