+++ src/screens/Ticketing/TicketDetails/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState} from 'react';
 import {NavigationProp, RouteProp} from '@react-navigation/native';
 import {RootStackParamList} from '../../../navigation';
 import Header from '../../../ScreenHeader';
@@ -7,6 +7,11 @@ import ThemeIcon from '../../../components/theme-icon';
 import {Close} from '../../../assets/svg/icons/actions';
 import ThemeText from '../../../components/text';
 import {SafeAreaView} from 'react-native-safe-area-context';
+import DetailedTicket from '../Ticket/DetailedTicket';
+import useInterval from '../../../utils/use-interval';
+import {useTicketState} from '../../../TicketContext';
+import {View} from 'react-native';
+import {ScrollView} from 'react-native-gesture-handler';
 
 export type TicketDetailsRouteParams = {
   orderId: string;
@@ -28,6 +33,12 @@ type Props = {
 
 export default function TicketDetails({navigation, route}: Props) {
   const styles = useStyles();
+  const [now, setNow] = useState<number>(Date.now());
+  useInterval(() => setNow(Date.now()), 2500);
+  const {fareContracts} = useTicketState();
+  const fc = fareContracts?.find(
+    (fc) => fc.order_id === route?.params?.orderId,
+  );
 
   return (
     <SafeAreaView style={styles.container}>
@@ -42,7 +53,9 @@ export default function TicketDetails({navigation, route}: Props) {
         title="Reisedetaljer"
         style={styles.header}
       />
-      <ThemeText>{route.params.orderId}</ThemeText>
+      <View style={styles.content}>
+        {fc && <DetailedTicket fareContract={fc} now={now} />}
+      </View>
     </SafeAreaView>
   );
 }
@@ -54,5 +67,8 @@ const useStyles = StyleSheet.createThemeHook((theme) => ({
   container: {
     flex: 1,
     backgroundColor: theme.background.level2,
+  },
+  content: {
+    padding: theme.spacings.medium,
   },
 }));
